+++ nengo_bones/scripts/__init__.py
@@ -1,0 +1,10 @@
+"""Scripts for managing repositories that use NengoBones."""
+
+from nengo_bones.scripts import (
+    base,
+    check_bones,
+    check_deploy,
+    format_notebook,
+    generate_bones,
+    pr_number,
+)