+++ nengo_bones/scripts/base.py
@@ -1,0 +1,8 @@
+"""Base command group that all scripts should import and use."""
+
+import click
+
+
+@click.group()
+def bones():
+    """See below for all commands provided by `bones`."""